move non-generic stuff to RandomAccessStore
